feat: revamp settings and catalog management
